requirements.txt
cassandra_driver==3.25.0
Flask==2.0.1
Flask_Login==0.5.0
Flask_Uploads==0.2.1
Flask_WTF==0.15.1
numpy==1.21.3
pdfkit==0.6.1
torch==1.10.0
torchvision==0.11.1
Werkzeug==2.0.1
WTForms==2.3.3
